fix(tests): move enterprise flags tests from ce/ to ee/ directory
Address PR review comments:
- Move authenticated theme and custom domain tests to ee/flags/ since
  they depend on enterprise flags hooks and customAppearanceEnabled
- Keep only the unauthenticated test in ce/flags/
- Remove unused TestContext import
